Windows warning fix.
svn path=/trunk/libavg/; revision=4988

+++ src/player/BlurFXNode.h
@@ -46,7 +46,7 @@ private:
 
     GPUBlurFilterPtr m_pFilter;
 
-    float m_StdDev;
+    double m_StdDev;
 };
 
 typedef boost::shared_ptr<BlurFXNode> BlurFXNodePtr;
